+++ test/api.jl
@@ -939,56 +939,6 @@ end
     end
 end
 
-@testset "Show" verbose = true begin
-    T = Float64
-
-    jd₀ = date_to_jd(2023, 1, 1, 0, 0, 0)
-    dt₀ = date_to_jd(2023, 1, 1, 0, 0, 0) |> julian2datetime
-    orb = KeplerianElements(
-        jd₀,
-        T(8000e3),
-        T(0.015),
-        T(28.5) |> deg2rad,
-        T(100) |> deg2rad,
-        T(200) |> deg2rad,
-        T(45) |> deg2rad,
-    )
-
-    # == J2 Orbit Propagator ===============================================================
-
-    orbp = Propagators.init(Val(:J2), orb)
-
-    expected = "J2 Orbit Propagator (Epoch = $(string(dt₀)), Δt = 0.0 s)"
-    result = sprint(show, orbp)
-    @test result == expected
-
-    expected = """
-        OrbitPropagatorJ2{Float64, Float64}:
-           Propagator name : J2 Orbit Propagator
-          Propagator epoch : $(string(dt₀))
-          Last propagation : $(string(dt₀))"""
-    result = sprint(show, MIME("text/plain"), orbp)
-
-    @test result == expected
-
-    # == J4 Orbit Propagator ===============================================================
-
-    orbp = Propagators.init(Val(:J4), orb)
-
-    expected = "J4 Orbit Propagator (Epoch = $(string(dt₀)), Δt = 0.0 s)"
-    result = sprint(show, orbp)
-    @test result == expected
-
-    expected = """
-        OrbitPropagatorJ4{Float64, Float64}:
-           Propagator name : J4 Orbit Propagator
-          Propagator epoch : $(string(dt₀))
-          Last propagation : $(string(dt₀))"""
-    result = sprint(show, MIME("text/plain"), orbp)
-
-    @test result == expected
-end
-
 @testset "Default Functions in the API" begin
     orbp = DummyPropagator{Float64, Float64}()
     @test Propagators.name(orbp) == "DummyPropagator{Float64, Float64}"